DefaultPrinter resets itself before building.
In certain situations where I was using a new instance of DefaultPrinter
on every test run, the DP was not resetting and removing the previous
'tests' div.
I'm not sure if maybe we should be getting runner.sections.run
to get the DP to reset (when at level1) instead of doing it here.