Added ability to isolate one or more tests by setting runner.only.
This might be useful when trying to debug a particular test.
Currently the default printer will print out all the sections even if they don't
contain tests in runner.only.
We either let this be; or get printer to only print the tests in
runner.only, or only print the sections which contain a test or who
have descendent subsections that do.